requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiofiles=23.2.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
attrs=23.2.0=pypi_0
ca-certificates=2023.12.12=haa95532_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
docopt=0.6.2=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.108.0=pypi_0
ffmpy=0.3.1=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.47.0=pypi_0
fsspec=2023.12.2=pypi_0
gradio=4.12.0=pypi_0
gradio-client=0.8.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.1=pypi_0
idna=3.6=pypi_0
importlib-resources=6.1.1=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
opencv-python=4.9.0.80=pypi_0
openssl=3.0.12=h2bbff1b_0
orjson=3.9.10=pypi_0
pandas=2.1.4=pypi_0
pillow=10.2.0=pypi_0
pip=23.3.1=py39haa95532_0
pipreqs=0.4.13=pypi_0
psutil=5.9.7=pypi_0
py-cpuinfo=9.0.0=pypi_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.6=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
referencing=0.32.0=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.16.2=pypi_0
scipy=1.11.4=pypi_0
seaborn=0.13.1=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=68.2.2=py39haa95532_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
starlette=0.32.0.post1=pypi_0
sympy=1.12=pypi_0
thop=0.1.1-2209072238=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.0=pypi_0
torch=2.1.2+cu118=pypi_0
torchaudio=2.1.2+cu118=pypi_0
torchvision=0.16.2+cu118=pypi_0
tqdm=4.66.1=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
ultralytics=8.0.234=pypi_0
urllib3=2.1.0=pypi_0
uvicorn=0.25.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
websockets=11.0.3=pypi_0
wheel=0.41.2=py39haa95532_0
yarg=0.1.9=pypi_0
zipp=3.17.0=pypi_0
